+++ index.js
@@ -1,27 +1,15 @@
 "use strict";
 
-const reactPlugin = require("eslint-plugin-react");
-const reactRules = require("./react-rules");
 const baseConfig = require("./base-config");
+const jsxConfig = require("./jsx");
+const tsConfig = require("./ts");
+const tsxConfig = require("./tsx");
 const testConfig = require("./test");
 
 module.exports = [
   baseConfig,
-  {
-    ...baseConfig,
-    files: [ "**/*.jsx" ],
-    languageOptions: {
-      ...baseConfig.languageOptions,
-      parserOptions: { ecmaFeatures: { jsx: true } },
-    },
-    plugins: { ...baseConfig.plugins, react: reactPlugin },
-    rules: { ...baseConfig.rules, ...reactRules },
-  },
-  {
-    ...baseConfig,
-    languageOptions: { ...baseConfig.languageOptions, ...testConfig.languageOptions },
-    plugins: { ...baseConfig.plugins, ...testConfig.plugins },
-    rules: { ...baseConfig.rules, ...testConfig.rules },
-    files: [ "test/**/*.js" ],
-  },
+  jsxConfig,
+  tsConfig,
+  tsxConfig,
+  testConfig,
 ];