+++ test.js
@@ -2,6 +2,7 @@
 
 const noOnlyTests = require("eslint-plugin-no-only-tests");
 const chaiFriendly = require("eslint-plugin-chai-friendly");
+const baseConfig = require("./base-config");
 
 const mochaCakes2Globals = {
   And: "readonly",
@@ -42,12 +43,16 @@ const mochaCakes2Rules = {
 };
 
 module.exports = {
+  ...baseConfig,
   languageOptions: { globals: { ...mochaCakes2Globals } },
+  files: [ "**/test/**/*.js" ],
   plugins: {
+    ...baseConfig.plugins,
     "no-only-tests": noOnlyTests,
     "chai-friendly": chaiFriendly,
   },
   rules: {
+    ...baseConfig.rules,
     ...mochaCakes2Rules,
     // no only in tests
     "no-only-tests/no-only-tests": [