+++ ts.js
@@ -1,16 +1,15 @@
 "use strict";
 
-const baseConfig = require(".");
+const baseConfig = require("./base-config");
 const typescriptRules = require("./typescript-rules");
 const typeScriptPlugin = require("@typescript-eslint/eslint-plugin");
 const typeScriptParser = require("@typescript-eslint/parser");
 const stylistic = require("@stylistic/eslint-plugin");
 const bnTypescriptRules = require("@bonniernews/eslint-plugin-typescript-rules");
-const reactPlugin = require("eslint-plugin-react");
-const reactRules = require("./react-rules");
 
 const typescriptBase = {
   plugins: {
+    ...baseConfig.plugins,
     "@typescript-eslint": typeScriptPlugin,
     "@stylistic": stylistic,
     "@bonniernews/typescript-rules": bnTypescriptRules,
@@ -20,18 +19,8 @@ const typescriptBase = {
     parser: typeScriptParser,
   },
   files: [ "**/*.ts" ],
-  rules: typescriptRules,
+  rules: { ...baseConfig.rules, ...typescriptRules },
   settings: { "import/resolver": { node: { extensions: [ ".ts", ".js" ] } } },
 };
 
-module.exports = [
-  ...baseConfig,
-  typescriptBase,
-  {
-    ...typescriptBase,
-    plugins: { ...typescriptBase.plugins, react: reactPlugin },
-    files: [ "**/*.tsx" ],
-    settings: { "import/resolver": { node: { extensions: [ ".ts", ".js", ".tsx", ".jsx" ] } } },
-    rules: { ...typescriptBase.rules, ...reactRules },
-  },
-];
+module.exports = typescriptBase;