+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## 2.0.0
 
 - Breaking change: switch to eslint 8 to eslint 9 and adapt to the new flat file config format.
-- Breaking change: removed `all`, `react`, `test` and `typescript-react` configs. Replaced with the default config (js) and
-  just one for typescript `typescript` which includes react and test rules.
+- Breaking change: removed `all`, `test` and `typescript-react` configs. Just importing the index config
+  will enable all configs. For just the specific configs import `@bonniernews/eslint-config/js`,
+  `@bonniernews/eslint-config/ts`, `@bonniernews/eslint-config/jsx`, `@bonniernews/eslint-config/test`
+  or `@bonniernews/eslint-config/tsx`.
 
 ## 1.2.0